+++ jest.config.js
@@ -1,0 +1,6 @@
+// Needed because of https://github.com/facebook/jest/issues/9771
+module.exports = {
+    'moduleNameMapper': {
+        'eslint/use-at-your-own-risk': 'eslint/lib/unsupported-api.js'
+      }
+}