+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to my ESLint config will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### BREAKING CHANGED
+- Changed rule 'camelCase' to '@typescript-eslint/naming-convention'
+
+### Added
+- Support for eslint@8
+- Support for eslint-plugin-sonarjs@0.11
+- Support for eslint-plugin-unicorn@39
+
 ## [5.2.0] - 2021-11-08
 ### Added
 - Support for eslint-plugin-unicorn@38